+++ languages/logo/logo.lang.revisitor/revisitor-comp/LogoProgram_revisitor/impl/operation/impl/MultOpImpl.java
@@ -39,6 +39,7 @@ import LogoProgram_revisitor.impl.operation.TurtleOp;
 import LogoProgram_revisitor.impl.operation.UnaryExpressionOp;
 import LogoProgram_revisitor.impl.operation.VariableOp;
 import LogoProgram_revisitor.impl.operation.WhileOp;
+import java.lang.Double;
 import kmLogo.Expression;
 import kmLogo.Mult;
 import kmLogo.Turtle;
@@ -58,7 +59,7 @@ public class MultOpImpl extends BinaryExpOpImpl implements MultOp {
 
   public double eval(Turtle turtle) {
     double result;
-    result = (rev.$((Expression)this.obj.getLhs()).eval(((Turtle) (turtle)))) * (rev.$((Expression)this.obj.getRhs()).eval(((Turtle) (turtle))));
+    result = ((Double) ((rev.$((Expression)this.obj.getLhs()).eval(((Turtle) (turtle)))) * (rev.$((Expression)this.obj.getRhs()).eval(((Turtle) (turtle))))));
     return result;
   }
 }
