+++ languages/logo/logo.lang.revisitor/revisitor-comp/LogoProgram_revisitor/impl/operation/impl/SinOpImpl.java
@@ -59,7 +59,7 @@ public class SinOpImpl extends UnaryExpressionOpImpl implements SinOp {
 
   public double eval(Turtle turtle) {
     double result;
-    result = TrigoServices.sinus((double) (rev.$((Expression)this.obj.getExpression()).eval(((Turtle) (turtle)))));
+    result = ((double) (TrigoServices.sinus((double) (rev.$((Expression)this.obj.getExpression()).eval(((Turtle) (turtle)))))));
     return result;
   }
 }
