+++ languages/logo/logo.lang.switch/switch-comp/emfswitch/emfswitch/operation/EqualsOperation.java
@@ -1,6 +1,7 @@
 package emfswitch.emfswitch.operation;
 
 import emfswitch.emfswitch.EmfswitchSwitchImplementation;
+import java.lang.Double;
 import kmLogo.Equals;
 import kmLogo.Turtle;
 import org.eclipse.emf.ecoretools.ale.compiler.lib.EqualService;
@@ -19,10 +20,10 @@ public class EqualsOperation extends BinaryExpOperation {
 	public double eval(Turtle turtle) {
 		double result;
 		if(EqualService.equals((((ExpressionOperation) emfswitch.doSwitch(this.it.getLhs())).eval((Turtle) (turtle))), (((ExpressionOperation) emfswitch.doSwitch(this.it.getRhs())).eval((Turtle) (turtle))))) {
-			result = 1.0;
+			result = ((Double) (1.0));
 		}
 		else {
-			result = 0.0;
+			result = ((Double) (0.0));
 		}
 		return result;
 	}
