+++ languages/logo/logo.lang.switch/switch-comp/emfswitch/emfswitch/operation/IfOperation.java
@@ -1,6 +1,7 @@
 package emfswitch.emfswitch.operation;
 
 import emfswitch.emfswitch.EmfswitchSwitchImplementation;
+import java.lang.Double;
 import kmLogo.If;
 import kmLogo.Turtle;
 import org.eclipse.emf.ecoretools.ale.compiler.lib.EqualService;
@@ -19,14 +20,14 @@ public class IfOperation extends ControlStructureOperation {
 	public double eval(Turtle turtle) {
 		double result;
 		if(!EqualService.equals((((ExpressionOperation) emfswitch.doSwitch(this.it.getCondition())).eval((Turtle) (turtle))), (0.0))) {
-			result = ((BlockOperation) emfswitch.doSwitch(this.it.getThenPart())).eval((Turtle) (turtle));
+			result = ((double) (((BlockOperation) emfswitch.doSwitch(this.it.getThenPart())).eval((Turtle) (turtle))));
 		}
 		else {
 			if(!EqualService.equals((this.it.getElsePart()), (null))) {
-				result = ((BlockOperation) emfswitch.doSwitch(this.it.getElsePart())).eval((Turtle) (turtle));
+				result = ((double) (((BlockOperation) emfswitch.doSwitch(this.it.getElsePart())).eval((Turtle) (turtle))));
 			}
 			else {
-				result = 0.0;
+				result = ((Double) (0.0));
 			}
 		}
 		return result;
