+++ languages/logo/logo.lang.switch/switch-comp/emfswitch/emfswitch/operation/ParameterCallOperation.java
@@ -1,6 +1,7 @@
 package emfswitch.emfswitch.operation;
 
 import emfswitch.emfswitch.EmfswitchSwitchImplementation;
+import java.lang.Double;
 import kmLogo.ParameterCall;
 import kmLogo.StackFrame;
 import kmLogo.Turtle;
@@ -20,11 +21,11 @@ public class ParameterCallOperation extends ExpressionOperation {
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = 0.0;
+		result = ((Double) (0.0));
 		for (StackFrame frame: turtle.getCallStack().getFrames()) {
 			for (Variable var: frame.getVariables()) {
 				if(EqualService.equals((var.getName()), (this.it.getParameter().getName()))) {
-					result = var.getValue();
+					result = ((double) (var.getValue()));
 				}
 			}
 		}
