+++ languages/logo/logo.lang.switch/switch-comp/emfswitch/emfswitch/operation/ProcCallOperation.java
@@ -1,9 +1,8 @@
 package emfswitch.emfswitch.operation;
 
-import org.eclipse.emf.ecoretools.ale.compiler.lib.CollectionService;
-import org.eclipse.emf.ecoretools.ale.compiler.lib.LogService;
-
 import emfswitch.emfswitch.EmfswitchSwitchImplementation;
+import java.lang.Double;
+import java.lang.Integer;
 import kmLogo.Expression;
 import kmLogo.KmLogoFactory;
 import kmLogo.ProcCall;
@@ -11,6 +10,8 @@ import kmLogo.ProcDeclaration;
 import kmLogo.StackFrame;
 import kmLogo.Turtle;
 import kmLogo.Variable;
+import org.eclipse.emf.ecoretools.ale.compiler.lib.CollectionService;
+import org.eclipse.emf.ecoretools.ale.compiler.lib.LogService;
 
 public class ProcCallOperation extends ExpressionOperation {
 	private final ProcCall it;
@@ -33,10 +34,10 @@ public class ProcCallOperation extends ExpressionOperation {
 			newVar.setName(CollectionService.get(this.it.getDeclaration().getArgs(), i).getName());
 			newVar.setValue(((ExpressionOperation) emfswitch.doSwitch(exp)).eval((Turtle) (turtle)));
 			newFrame.getVariables().add(newVar);
-			i = (i) + (1);
+			i = ((Integer) ((i) + (1)));
 		}
 		turtle.getCallStack().getFrames().add(newFrame);
-		result = 0.0;
+		result = ((Double) (0.0));
 		if(this.it.getDeclaration() instanceof ProcDeclaration) {
 			ProcDeclaration decl = ((ProcDeclaration) (this.it.getDeclaration()));
 			((ProcDeclarationOperation) emfswitch.doSwitch(decl)).deval((Turtle) (turtle));
