+++ languages/logo/logo.lang.switch/switch-comp/emfswitch/emfswitch/operation/TurtleOperation.java
@@ -1,12 +1,12 @@
 package emfswitch.emfswitch.operation;
 
-import org.eclipse.emf.ecoretools.ale.core.interpreter.services.TrigoServices;
-
 import emfswitch.emfswitch.EmfswitchSwitchImplementation;
+import java.lang.Double;
 import kmLogo.KmLogoFactory;
 import kmLogo.Point;
 import kmLogo.Segment;
 import kmLogo.Turtle;
+import org.eclipse.emf.ecoretools.ale.core.interpreter.services.TrigoServices;
 
 public class TurtleOperation {
 	private final Turtle it;
@@ -43,12 +43,12 @@ public class TurtleOperation {
 	public void rotate(double angle) {
 		double newAngle = ((double) ((this.it.getHeading()) + (angle)));
 		if((newAngle) > (360.0)) {
-			newAngle = (newAngle) - (360.0);
+			newAngle = ((Double) ((newAngle) - (360.0)));
 			this.it.setHeading(newAngle);
 		}
 		else {
 			if((newAngle) < (0.0)) {
-				newAngle = (360.0) + (newAngle);
+				newAngle = ((Double) ((360.0) + (newAngle)));
 				this.it.setHeading(newAngle);
 			}
 			else {
